+++ packages/hdl-panels/src/panels/panelsForCoMLOpsPromptEngineeringEx2/components/AnnotationInputForm.stories.tsx
@@ -1,5 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/react";
 
+import { tagOptionsForEachTagType } from "src/panels/panelsForCoMLOpsPromptEngineeringEx2/_hardCordingValue";
 import { AnnotationInputForm } from "./AnnotationInputForm";
 
 const meta: Meta<typeof AnnotationInputForm> = {
@@ -13,14 +14,7 @@ type Story = StoryObj<typeof AnnotationInputForm>;
 export const Basic: Story = {
   render: () => {
     return (
-      <AnnotationInputForm
-        onSave={() => {}}
-        tagOptions={[
-          { label: "test1", value: "test1" },
-          { label: "test2", value: "test2" },
-        ]}
-        tagType="test1"
-      />
+      <AnnotationInputForm onSave={() => {}} tagOptionsForEachTagType={tagOptionsForEachTagType} />
     );
   },
 };
