+++ packages/hdl-panels/src/panels/panelsForCoMLOpsPromptEngineeringEx2/components/AnnotationInputForm.tsx
@@ -9,7 +9,7 @@ import {
 } from "@mui/material";
 import { Box } from "@mui/system";
 import { useState } from "react";
-import { TagOptionsForEachTagType, UserInputAnnotation } from "./types";
+import { TagOptionsForEachTagType, UserInputAnnotation } from "../types";
 
 export const AnnotationInputForm = ({
   defaultValues,
