+++ packages/hdl-panels/src/panels/panelsForCoMLOpsPromptEngineeringEx2/components/AnnotationTable.stories.tsx
